simplify the client cache
