+++ pkg/kubectl/cmd/util/BUILD
@@ -4,7 +4,6 @@ go_library(
     name = "go_default_library",
     srcs = [
         "cached_discovery.go",
-        "clientcache.go",
         "factory.go",
         "factory_builder.go",
         "factory_client_access.go",
